+++ src/components/groups/custom/cos/plans/PlansSection.tsx
@@ -18,7 +18,11 @@ export default async function PlansSection({ groupID }: { groupID: string }) {
 
     return (
       <div className="w-full">
-        <h1 className="text-lg font-bold text-custom-dark-text">Planned COS</h1>
+        <div className="flex items-center justify-between gap-2 w-full">
+          <h1 className="text-lg font-bold text-custom-dark-text">
+            Planned COS
+          </h1>
+        </div>
         <HRow className="mb-2" />
         {Object.keys(cosData).length === 0 ? (
           <div className="min-h-[10vh] bg-white grid place-items-center p-2 rounded-md">
